Use libxmljs for XmlLoader XML parsing and updating
The *Loader classes in module-loader expose an update() method for updating
the internal representation of a parsed XML or YAML file and returning the
result as an XML or YAML string. xmldoc, which the XmlLoader class uses to
parse XML docs, doesn't appear to be designed for manipulating parsed XML,
but only for reading it, and its XMLDocument and XMLElement classes don't
provide any public interface for updating their contents. This is currently
worked around by manipulating the 'val' property of xmldoc XMLElement
objects of interest.[1]

This commit updates XmlLoader to use libxmljs, which provides a proper
interface for manipulating XML Document objects. This allows us to move
off our specialized xmldoc fork and makes us more robust going forward.

Phab: https://phabricator.wikimedia.org/T177019